# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthVerifyBody {
  /** Firebase ID token from client-side authentication */
  idToken: string;
}

export interface AuthVerifyResponse {
  /** JWT session token for subsequent API calls */
  sessionToken: string;
  /** Current credit balance */
  credits: number;
  /** User display name (email or masked phone) */
  displayName: string;
}

export interface GenerateReportBody {
  /** Name of the date partner */
  partnerName: string;
  /** Type of date occasion (e.g., coffee, dinner, movie) */
  occasion: string;
  /** Any known details about the partner */
  knownDetails?: string;
  /** The vibe or mood you want to set */
  vibe?: string;
}

export interface ReportSection {
  title: string;
  content: string;
  items?: string[];
  /** Index (0-based) of the single strongest item in this section */
  recommendedIndex?: number;
}

export interface IntelligenceReport {
  partnerName: string;
  innerGame: ReportSection;
  openingGambit: ReportSection;
  iqQuestions: ReportSection;
  conversationClosers: ReportSection;
}

export type GenerateReportResponse = IntelligenceReport & {
  /** Credits left after this report was generated */
  creditsRemaining: number;
};

export interface SubmitUtrBody {
  /**
   * UPI Transaction Reference (UTR) — 12–50 alphanumeric characters
   * @minLength 12
   * @maxLength 50
   */
  utr: string;
  /** Unique session ID for the current report */
  reportSession: string;
}

export interface SubmitUtrResponse {
  ok: boolean;
  /** HMAC token confirming UTR submission was recorded server-side */
  token: string;
  reportSession: string;
  /** Credit balance after the top-up */
  creditsRemaining: number;
}

export interface PaymentStatusResponse {
  approved: boolean;
  /** Unix timestamp (ms) when submission was approved */
  approvedAt?: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}