# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Firebase auth verification
  - name: report
    description: Intelligence report generation
  - name: payment
    description: UPI payment submission and status
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/verify:
    post:
      operationId: verifyAuth
      tags: [auth]
      summary: Verify Firebase ID token and create/find user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyBody"
      responses:
        "200":
          description: Auth verified, session token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Auth failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /report/generate:
    post:
      operationId: generateReport
      tags: [report]
      summary: Generate a date intelligence report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportBody"
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateReportResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No credits remaining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment/submit-utr:
    post:
      operationId: submitUtr
      tags: [payment]
      summary: Submit a UPI UTR for payment verification
      description: Validates UTR format, records the submission server-side, credits 5 reports to the user, and returns a session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitUtrBody"
      responses:
        "200":
          description: UTR accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitUtrResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment/status/{sessionId}:
    get:
      operationId: getPaymentStatus
      tags: [payment]
      summary: Check payment status for a report session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthVerifyBody:
      type: object
      properties:
        idToken:
          type: string
          description: Firebase ID token from client-side authentication
      required:
        - idToken

    AuthVerifyResponse:
      type: object
      properties:
        sessionToken:
          type: string
          description: JWT session token for subsequent API calls
        credits:
          type: integer
          description: Current credit balance
        displayName:
          type: string
          description: User display name (email or masked phone)
      required:
        - sessionToken
        - credits
        - displayName

    GenerateReportBody:
      type: object
      properties:
        partnerName:
          type: string
          description: Name of the date partner
        occasion:
          type: string
          description: Type of date occasion (e.g., coffee, dinner, movie)
        knownDetails:
          type: string
          description: Any known details about the partner
        vibe:
          type: string
          description: The vibe or mood you want to set
      required:
        - partnerName
        - occasion

    ReportSection:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        items:
          type: array
          items:
            type: string
        recommendedIndex:
          type: integer
          description: Index (0-based) of the single strongest item in this section
      required:
        - title
        - content

    IntelligenceReport:
      type: object
      properties:
        partnerName:
          type: string
        innerGame:
          $ref: "#/components/schemas/ReportSection"
        openingGambit:
          $ref: "#/components/schemas/ReportSection"
        iqQuestions:
          $ref: "#/components/schemas/ReportSection"
        conversationClosers:
          $ref: "#/components/schemas/ReportSection"
      required:
        - partnerName
        - innerGame
        - openingGambit
        - iqQuestions
        - conversationClosers

    GenerateReportResponse:
      allOf:
        - $ref: "#/components/schemas/IntelligenceReport"
        - type: object
          properties:
            creditsRemaining:
              type: integer
              description: Credits left after this report was generated
          required:
            - creditsRemaining

    SubmitUtrBody:
      type: object
      properties:
        utr:
          type: string
          minLength: 12
          maxLength: 50
          description: UPI Transaction Reference (UTR) — 12–50 alphanumeric characters
        reportSession:
          type: string
          description: Unique session ID for the current report
      required:
        - utr
        - reportSession

    SubmitUtrResponse:
      type: object
      properties:
        ok:
          type: boolean
        token:
          type: string
          description: HMAC token confirming UTR submission was recorded server-side
        reportSession:
          type: string
        creditsRemaining:
          type: integer
          description: Credit balance after the top-up
      required:
        - ok
        - token
        - reportSession
        - creditsRemaining

    PaymentStatusResponse:
      type: object
      properties:
        approved:
          type: boolean
        approvedAt:
          type: number
          description: Unix timestamp (ms) when submission was approved
      required:
        - approved

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message