# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Verify Firebase ID token and create/find user
 */
export const VerifyAuthBody = zod.object({
  idToken: zod
    .string()
    .describe("Firebase ID token from client-side authentication"),
});

export const VerifyAuthResponse = zod.object({
  sessionToken: zod
    .string()
    .describe("JWT session token for subsequent API calls"),
  credits: zod.number().describe("Current credit balance"),
  displayName: zod
    .string()
    .describe("User display name (email or masked phone)"),
});

/**
 * @summary Generate a date intelligence report
 */
export const GenerateReportBody = zod.object({
  partnerName: zod.string().describe("Name of the date partner"),
  occasion: zod
    .string()
    .describe("Type of date occasion (e.g., coffee, dinner, movie)"),
  knownDetails: zod
    .string()
    .optional()
    .describe("Any known details about the partner"),
  vibe: zod.string().optional().describe("The vibe or mood you want to set"),
});

export const GenerateReportResponse = zod
  .object({
    partnerName: zod.string(),
    innerGame: zod.object({
      title: zod.string(),
      content: zod.string(),
      items: zod.array(zod.string()).optional(),
      recommendedIndex: zod
        .number()
        .optional()
        .describe(
          "Index (0-based) of the single strongest item in this section",
        ),
    }),
    openingGambit: zod.object({
      title: zod.string(),
      content: zod.string(),
      items: zod.array(zod.string()).optional(),
      recommendedIndex: zod
        .number()
        .optional()
        .describe(
          "Index (0-based) of the single strongest item in this section",
        ),
    }),
    iqQuestions: zod.object({
      title: zod.string(),
      content: zod.string(),
      items: zod.array(zod.string()).optional(),
      recommendedIndex: zod
        .number()
        .optional()
        .describe(
          "Index (0-based) of the single strongest item in this section",
        ),
    }),
    conversationClosers: zod.object({
      title: zod.string(),
      content: zod.string(),
      items: zod.array(zod.string()).optional(),
      recommendedIndex: zod
        .number()
        .optional()
        .describe(
          "Index (0-based) of the single strongest item in this section",
        ),
    }),
  })
  .and(
    zod.object({
      creditsRemaining: zod
        .number()
        .describe("Credits left after this report was generated"),
    }),
  );

/**
 * Validates UTR format, records the submission server-side, credits 5 reports to the user, and returns a session token.
 * @summary Submit a UPI UTR for payment verification
 */
export const submitUtrBodyUtrMin = 12;
export const submitUtrBodyUtrMax = 50;

export const SubmitUtrBody = zod.object({
  utr: zod
    .string()
    .min(submitUtrBodyUtrMin)
    .max(submitUtrBodyUtrMax)
    .describe(
      "UPI Transaction Reference (UTR) — 12–50 alphanumeric characters",
    ),
  reportSession: zod
    .string()
    .describe("Unique session ID for the current report"),
});

export const SubmitUtrResponse = zod.object({
  ok: zod.boolean(),
  token: zod
    .string()
    .describe("HMAC token confirming UTR submission was recorded server-side"),
  reportSession: zod.string(),
  creditsRemaining: zod.number().describe("Credit balance after the top-up"),
});

/**
 * @summary Check payment status for a report session
 */
export const GetPaymentStatusParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetPaymentStatusResponse = zod.object({
  approved: zod.boolean(),
  approvedAt: zod
    .number()
    .optional()
    .describe("Unix timestamp (ms) when submission was approved"),
});